pyproject.toml
[project]
name = "data-science"
version = "0.1.0"
description = "A Data Science agent for automated SQL generation, data analysis, and telemetry-instrumented workflows."
authors = [
    { name = "Google LLC" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=1.14",
    "immutabledict>=4.2.1",
    "sqlglot>=26.10.1",
    "db-dtypes>=1.4.2",
    "regex>=2024.11.6",
    "tabulate>=0.9.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "absl-py>=2.2.2",
    "pydantic>=2.11.3",
    "pandas>=2.3.0",
    "numpy>=2.3.1",
    "toolbox-core>=0.3.0",
    "pg8000>=1.31.2",
    "psycopg2-binary>=2.9.0",
    "geopandas>=1.0.0",
    "geoalchemy2>=0.15.0",
    "sqlalchemy>=2.0.0",
    "google-cloud-discoveryengine>=0.11.0",
    "requests>=2.31.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
]

[dependency-groups]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.14",
    "ruff>=0.4.6",
]

[tool.ruff]
# Points back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint]
ignore = ["C901", "PLR0911", "PLR0912", "PLW0603"]

[tool.ruff.lint.isort]
# Ensures the data science source is treated as first-party
known-first-party = ["data_science"]

[tool.hatch.build.targets.wheel]
# Maps the package source correctly for Hatchling
packages = ["data_science"]

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"